Renamed CreateRoutine to CreateSession2, copied old CreateRoutine
I've modified the CreateRoutine class to be used as a CreateSession instead. Renamed the file and copied the old CreateRoutine as to not loose the code. Please note that I have not written it
