+++ TrainingApp/app/src/main/java/com/traininapp/View/UpcomingFragment.java
@@ -15,11 +15,9 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.traininapp.MainActivity;
 import com.traininapp.Model.Planning.Session;
 import com.traininapp.R;
-import com.traininapp.viewModel.CreateRoutine;
-import com.traininapp.viewModel.CreateSession;
+import com.traininapp.viewModel.CreateSession2;
 import com.traininapp.viewModel.UpcomingSessionsViewModel;
 
 import java.time.LocalDate;
@@ -109,7 +107,7 @@ public class UpcomingFragment extends Fragment {
     public void onAddSessionClick(){
 
         // Creating and initializing the intent object
-        Intent intent = new Intent(getActivity(), CreateRoutine.class);
+        Intent intent = new Intent(getActivity(), CreateSession2.class);
 
         /*// Attaching the key value pair using putExtra to this intent
         intent.putExtra("VIEW_MODEL", upcomingSessionsViewModel);*/
